Move QuantMEMatcher
